Finally broke the class in two: addr and iface.
Replaced the is_computer_interface_address() with using the new and improved find_addr_interface() so we have additional info, not just true/false.
Added an is_default() function since that's is true when we setup an addr object. That way you can detect an uninitialized addr.
Updated the tests accordingly. They still need to to be enhanced to cover 100% of the code.