Added the ADDRESS_LOOKUP flag.
Removed the MULTI_ADDRESSES_COMMAS_AND_SPACES. Set both MULTI_ADDRESSES_COMMAS
and MULTI_ADDRESSES_SPACES instead.

Updated the test verifying that the COMMAS are exclusive between addresses
and ports.

Make use of the get_allow() everywhere so the implementation can have some
special cases if we'd like to.

Removed the useless 'struct' (not necessary in C++).

Changed the strategy to initialize the flags at the start: it's now done in
the constructor. This is much clearer and allows us to change the enum to
any order and it still works as expected.