Added support for an address validator.
The advgetopt library offers a way to add validators. These allow us to
check that a parameter is properly defined, that the value is considered
valid as per the specified validator.

The address validator makes sure that the libaddr can indeed parse that
IP address or convert a hostname to an IP address range.