Added to_addresses() to the addr_range class.
Added an out_of_range exception raised when the number of address is larger than the limit in to_addresses().
Added a size() function to the addr_range class.
Added is_lan()/is_wan() to check groups of IPs.
Added the ip_from_uint128() as the counter part of ip_to_uint128().
Documented the ability to bind to port 0 which auto-selects a port.
Added a non-const bind() call to auto-fill the port after the binding.
Added the +, -, +=, -=, ++, and -- operators to the addr object.
Added the CATCH_START_SECTION()/CATCH_END_SECTION() to all tests.
Test a few more of the new functions from the last few iterations (family, hostname, interface, mask).
Various clean ups.