Convert mask to just one number if possible.
I had a TODO comment about converting the /<mask> in a number if the mask
represents such (i.e. all 1s then all 0s from the most significant to the
least significant bits). Now this is implemented since I have the
get_mask_size() function for the purpose and it's very easy to use.