Fixed a memory leak in the std::ostream implementation.
The setaddrsep capture would allocate a new ostrem_info but not register the callback.
Made a few calls lighter in the ostream implementation.
Use get_port() everywhere in our addr implementation (it takes care of swapping the bytes automatically).
Added a get_str_port() because that's something we do often.
Added test to verify the ostream functions.