Change the ip to string code.
The ip to string code was very rigid. You could not just print out the port
or the mask. This newer version allows you to decide between the address,
port, and mask separately. Similarly, the brackets can be added to the
address or mask (actually, this is not totally true, but close). You can
also force the function to print the mask as an IP address (as expected
by the --mask option of the iptables command).