Fixed mask parsing.
The parsing would use the address is_ipv4() which could return true even when
the address was defined as an IPv6 address. That meant a mask of 96 would
fail since the maximum in IPv4 is 32. The CIDR for an IPv6, even if it
represents an IPv4, must now be an IPv6. This fails if the input address was
given as a domain name. In that case, the parser things it's IPv4 even if the
address is an IPv6 address. But you probably should not ever use a mask with
a domain name anyway.

Fixed a help message of the validate_ip tool.