Enhanced unit tests.
Found out how I could hit a few more lines of code while working on cluck and
decided to add that to a test here. This eliminates some LCOV_EXCL_LINE and
instead verifies those lines.

Also ran in a few errors where the newer versions of Linux return a slightly
different error code (errno of 22 instead of 2) in a few circumstances.

Got an error with the routes, if the destination is not 0.0.0.0 then the
gateway is not the default gateway so we should not mark it as found in
that case.