Keep CIDR as just a number even with IPv6.
To pass the default CIDR to parse_mask(), it would be written between
'[' and ']' when in IPv6. Now we add those only if the address includes
at least one ':'. This means we correctly handle the '[' and ']'
in the parse_mask() function.

Also updated a few tests to properly respect this new behavior.