Ignore duplicates when dealing with ranges.
If a range detects multiple protocols with the same IP address, the protocol
is dropped (the first is kept). In most likelihood, this will resolve issues
where you forgot to setup the protocol first.

Changed the comment implementation so if the '\n' is one of the separators,
then the entire rest of the line is considered a comment, whether other IP
separators apply, like the comma or space. This means those extra separators
do not apply within those comments.

Fixed many tests so they pass under Ubuntu 24.04 and also added some tests
for better coverage of the newer code.